Add an optimization to the spitfire compiler that allows for batch writing to the buffer.
This optimization avoids resizing the buffer(list) on _buffer_append. The new option, batch_buffer_writes, controls the optimization.

Further optimizations include exploring the If, For, Else, etc. nodes for buffer writes and only extending before that node if it has writes. I am not 100% sure if that is safe though.

[git-p4: change = 76616970]