Function macro parse selection
Function macros are currently parsed as rhs_expressions. This is because all statements that just contain an expression are implicitly turned into writes. Analyze placeholder substitution node is what performs this implicit transformation. This change allows for function macros to be parsed as statements. This is then checked by the placeholder substitution analyzer and if the result is a statement, there is no write added.

[git-p4: change = 78712962]