Baked Spitfire Optimization: Directly construct SantizedPlaceholder when function is SanitizedState.SANITIZED_STRING
Also do not sanitize function calls in #do statements.

This is safe because SanitizedState.STRING is when the function is a template method or a  template library function.

Both of cases return strings that are already escaped. Thus, this optimization is considered safe.

This change also renames the SanitizedState enum to SANITIZED_STRING, SANITIZED, UNKNOWN, UNSANITIZED, and NOT_OUTPUTTED for what we know about the function call.

[git-p4: change = 85034808]