Make all empty string literals in spitfire sanitized by default.
In order to facilitate a pattern where multiple strings of html are built up in a loop simultaneously, it is helpful to be able to make that string sanitized. Because the '+' operator returns a sanitized placeholder when both the left and right operands are sanitized, by providing all empty strings are sanitized by default, building up a sanitized string in spitfire from an empty start works as expected.

[git-p4: change = 98354502]